Tweaked enemy spawn system, added total resource score display

+++ Feast or Famine/Assets/Scripts/Tower Defense/Enemies/TDEnemy.cs
@@ -28,7 +28,7 @@ public class TDEnemy : MonoBehaviour
     {
         DamagePerAttack *= PlayerPrefs.GetInt("RoundScaling");
         health *= PlayerPrefs.GetInt("RoundScaling");
-        Speed *= PlayerPrefs.GetInt("RoundScaling");
+        Speed *= ((float)PlayerPrefs.GetInt("RoundScaling"));
     }
 
     void Update()
